+++ find.d.ts
@@ -1,7 +1,8 @@
 import type {ImmutableTree} from './types';
 
-export default function find<T, V = T>(
+export default function find<T, K = T, D = T>(
   tree: ImmutableTree<T> | null,
-  value: V,
-  cmp: (a: V, b: T) => number,
-): ImmutableTree<T> | null;
+  key: K,
+  cmp: (a: K, b: T) => number,
+  defaultValue: D,
+): T | D;