+++ findNext.mjs
@@ -5,15 +5,16 @@ import minNode from './minNode.mjs';
 import type {ImmutableTree} from './types.mjs';
 */
 
-export default function findNext/*:: <T, V = T> */(
+export default function findNext/*:: <T, K = T, D = T> */(
   tree/*: ImmutableTree<T> | null */,
-  value/*: V */,
-  cmp/*: (a: V, b: T) => number */,
-)/*: ImmutableTree<T> | null */ {
+  key/*: K */,
+  cmp/*: (a: K, b: T) => number */,
+  defaultValue/*: D */,
+)/*: T | D */ {
   let cursor = tree;
   let largerParent = null;
   while (cursor !== null) {
-    const order = cmp(value, cursor.value);
+    const order = cmp(key, cursor.value);
     if (order === 0) {
       break;
     } else if (order < 0) {
@@ -24,7 +25,10 @@ export default function findNext/*:: <T, V = T> */(
     }
   }
   if (cursor !== null && cursor.right !== null) {
-    return minNode(cursor.right);
+    return minNode(cursor.right).value;
   }
-  return largerParent;
+  if (largerParent !== null) {
+    return largerParent.value;
+  }
+  return defaultValue;
 }