+++ findPrev.mjs
@@ -5,15 +5,16 @@ import maxNode from './maxNode.mjs';
 import type {ImmutableTree} from './types.mjs';
 */
 
-export default function findPrev/*:: <T, V = T> */(
+export default function findPrev/*:: <T, K = T, D = T> */(
   tree/*: ImmutableTree<T> | null */,
-  value/*: V */,
-  cmp/*: (a: V, b: T) => number */,
-)/*: ImmutableTree<T> | null */ {
+  key/*: K */,
+  cmp/*: (a: K, b: T) => number */,
+  defaultValue/*: D */,
+)/*: T | D */ {
   let cursor = tree;
   let smallerParent = null;
   while (cursor !== null) {
-    const order = cmp(value, cursor.value);
+    const order = cmp(key, cursor.value);
     if (order === 0) {
       break;
     } else if (order < 0) {
@@ -24,7 +25,10 @@ export default function findPrev/*:: <T, V = T> */(
     }
   }
   if (cursor !== null && cursor.left !== null) {
-    return maxNode(cursor.left);
+    return maxNode(cursor.left).value;
   }
-  return smallerParent;
+  if (smallerParent !== null) {
+    return smallerParent.value;
+  }
+  return defaultValue;
 }